backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
stripe==7.8.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
phonenumbers==8.13.25
twilio==8.10.3
authlib==1.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
